requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.2.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.9.11
cffi==1.12.3
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
imageio==2.6.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vis==0.4.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
#mkl-fft==1.0.14
#mkl-random==1.1.0
#mkl-service==2.3.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
np-utils==0.5.11.1
numpy==1.16.5
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow>=7.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
pycparser==2.19
pydot==1.4.1
pydot-ng==2.0.0
pydotplus==2.0.2
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.10.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.15.4
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
#torch==1.2.0
#torchvision==0.5.0+cpu
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
